small directory names cleanup na organizing
