+++ src/backend/AuthService/Actions/UserSettingsActions/GetUserDetails/GetUserDetailsService.cs
@@ -1,0 +1,40 @@
+﻿using Core.Auth0;
+using Core.Database;
+using Core.Database.Models.Auth;
+using Core.Exceptions;
+using Microsoft.EntityFrameworkCore;
+using Microsoft.Extensions.Logging;
+using System.Linq.Dynamic.Core;
+
+namespace AuthService.Actions.UserSettingsActions.GetUserDetails
+{
+    public class GetUserDetailsService
+    {
+        private readonly AppDbContext _appDbContext;
+        private readonly ILogger<GetUserDetailsService> _logger;
+
+        public GetUserDetailsService(AppDbContext appDbContext, ILogger<GetUserDetailsService> logger)
+        {
+            _appDbContext = appDbContext;
+            _logger = logger;
+        }
+
+        public async Task<User> GetCurrentUser(CurrentUserInfo auth0UserInfo)
+        {
+            var userFromDatabase = await _appDbContext.Users
+                .Include(c => c.Auth0Users!)
+                .Include(c => c.Roles!).ThenInclude(c => c.Role!)
+                .Where(c => c.Auth0Users!.Any(s => s.Identifier == auth0UserInfo.Identifier))
+                .SingleOrDefaultAsync();
+
+            if (userFromDatabase == null)
+            {
+
+                _logger.LogError("User with email {email} was not found in table Auth0Users", auth0UserInfo.Email);
+                throw new HttpError500Exception("Sorry we couldn't find your user in database");
+            }
+
+            return userFromDatabase!;
+        }
+    }
+}
