+++ src/backend/AuthService/Actions/UserSettingsActions/PatchUserDetails/PatchUserDetailsPayload.cs
@@ -1,0 +1,8 @@
+﻿namespace AuthService.Actions.UserSettingsActions.PatchUserDetails
+{
+    public class PatchUserDetailsPayload
+    {
+        public string FirstName { get; set; } = default!;
+        public string LastName { get; set; } = default!;
+    }
+}
