+++ src/backend/AuthService/AuthServiceModule.cs
@@ -1,10 +1,10 @@
-﻿using AuthService.BusinessLogic.PostLogin;
-using AuthService.BusinessLogic.GetAuth0UsersByCurrentUser;
-using Autofac;
+﻿using Autofac;
 using Core.Autofac;
-using AuthService.BusinessLogic.PatchUserDetails;
-using AuthService.BusinessLogic.GetUserDetails;
 using AuthService.Services;
+using AuthService.Actions.AuthActions.PostLogin;
+using AuthService.Actions.UserSettingsActions.PatchUserDetails;
+using AuthService.Actions.UserSettingsActions.GetAuth0UsersByCurrentUser;
+using AuthService.Actions.UserSettingsActions.GetUserDetails;
 
 namespace AuthService
 {
