--- src/backend/AuthService/BusinessLogic/PatchUserDetails/PatchUserDetailsPayloadValidator.cs
@@ -1,13 +1,0 @@
-﻿using FluentValidation;
-
-namespace AuthService.BusinessLogic.PatchUserDetails
-{
-    public class PatchUserDetailsPayloadValidator : AbstractValidator<PatchUserDetailsPayload>
-    {
-        public PatchUserDetailsPayloadValidator()
-        {
-            RuleFor(c => c.FirstName).NotNull().Length(1, 50);
-            RuleFor(c => c.LastName).NotNull().Length(1, 50);
-        }
-    }
-}
